medical_ai_service/requirements.txt
Flask==2.3.3
torch==2.8.0
torchvision==0.23.0
Pillow==10.0.0
numpy==1.24.3
Werkzeug==2.3.7
